chore: cleanup mock files and move test files
