+++ packages/shared/lib/tests/__mocks__/api.mock.ts
@@ -1,7 +1,9 @@
 import { AccountId, AccountManagerOptions } from '@iota/wallet'
-import { IApi } from '../../core/profile-manager'
+
+import { IApi } from '@core/profile-manager'
+
 import { AccountMock } from './account.mock'
-import { ProfileManagerMock } from './profileManager.mock'
+import { ProfileManagerMock } from './profile-manager.mock'
 
 const api: IApi = {
     createAccountManager(_: AccountManagerOptions): ProfileManagerMock {
